pyproject.toml
[tool.poetry]
name = "crew"
version = "0.1.0"
description = "cool"
authors = ["Your Name <you@example.com>"]
license = "n"
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.5"
crewai = "^0.74.2"
crewai-tools = "^0.13.2"
langchain-google-genai = "^2.0.1"
load-dotenv = "^0.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
